pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "humand-sdk"
version = "0.1.0"
authors = [
    {name = "Humand Team", email = "support@humand.io"},
]
description = "Human-in-the-Loop Approval System for LangGraph"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["approval", "human-in-the-loop", "langgraph", "workflow", "ai", "automation"]
dependencies = [
    "requests>=2.25.0",
    "typing-extensions>=4.0.0; python_version<'3.8'",
]

[project.optional-dependencies]
langgraph = [
    "langgraph>=0.0.60",
    "langchain-core>=0.1.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]
examples = [
    "langgraph>=0.0.60",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/humand-io/humand-sdk"
Documentation = "https://docs.humand.io"
Repository = "https://github.com/humand-io/humand-sdk"
"Bug Reports" = "https://github.com/humand-io/humand-sdk/issues"

[project.scripts]
humand = "humand_sdk.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["humand_sdk*"]
exclude = ["tests*", "examples*"]

[tool.setuptools.package-data]
humand_sdk = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["humand_sdk"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
